Adjust the failing tests
The existing exception strings in the tests do not pass the
"domain_name in str(e)" condition found in many dns plugins,
when the new "_acme-challenge" base domain guess is executed.

This commit is quite dangerous - the necessary testing has only been
done on dnsmadeeasy, in their sandbox. Actually, dnsmadeeasy does not
allow domain names to start with "_acme-challenge". Therefore, what
matters is that in the common case the additional guess of the base
domain name does not do any harm.

There may well be other providers that return a different (and
unexpected) exception when asked about the non-existing _acme-challenge
zone.