dns_common_lexicon: deprecate the "domain" parameter for add/del_txt_record
There is no valid use case for it, and attempts to use it to find
the correct zone resulted only in bugs when _acme-challenge is a
separate zone. However, external plugins still use it, that's why
deprecation and not complete removal.

Implementation notes:

1. Tests cannot use `assertWarns()` because of Python 2.7. However,
the call to `warnings.catch_warnings(record=True)` hits a bug in mypy:
https://github.com/python/typeshed/issues/3463
That's why `assert w is not None`.

2. CI tests old plugins against new certbot. That's why both new
and deprecated calls to add/del_txt_record() are accepted as correct
in BaseLexiconAuthenticatorTest.